backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
requests>=2.32.2
numpy>=1.26.4
# ML Dependencies
torch>=2.4.1
transformers>=4.45.0
soundfile>=0.12.1
librosa>=0.10.1
scipy>=1.12.0
datasets>=2.21.0
accelerate>=0.28.0
supabase>=2.3.0
python-dotenv>=1.0.1
websockets>=12.0
pydub>=0.25.1
piper-tts>=1.2.0
pathvalidate>=3.0.0
